style: clean up browser and list keymaps
